// frontend/src/app/reducers/action-types.ts
export enum ActionType {
    FORM_IS_VALID = 'FORM_IS_VALID',
    FIRST_NAME_UPDATE = 'FIRST_NAME_UPDATE',
    LAST_NAME_UPDATE = 'LAST_NAME_UPDATE',
    EMAIL_UPDATE = 'EMAIL_UPDATE',
    PHONE_UPDATE = 'PHONE_UPDATE',
    MESSAGE_UPDATE = 'MESSAGE_UPDATE',
    POMSKY_NAME = 'POMSKY_NAME',
    SUBMISSION = 'SUBMISSION'
}